+++ app/src/main/java/woowacourse/payments/ui/addcard/AddCardActivity.kt
@@ -13,6 +13,7 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.ui.addcard.component.AddCardTopbar
+import woowacourse.payments.ui.addcard.component.AddCardVendorModal
 import woowacourse.payments.ui.addcard.model.VendorModalUiState
 import woowacourse.payments.ui.allcards.AllCardsActivity.Companion.CARD_INFO_KEY
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
@@ -45,18 +46,8 @@ class AddCardActivity : ComponentActivity() {
                                 .padding(horizontal = 24.dp)
                                 .fillMaxWidth(),
                         cardInfo = cardInfo,
+                        vendorModal = vendorModal,
                         onCardClick = { vendorModal.show() },
-                    )
-
-                    AddCardVendorModalScreen(
-                        vendorModalUiState = vendorModal,
-                        onDismissRequest = { vendorModal.hide() },
-                        onVendorItemClick = { vendor ->
-                            cardInfo.updateCardInfo(
-                                vendor = vendor,
-                            )
-                            vendorModal.hide()
-                        },
                     )
                 }
             }
